pesquisa em user-management e novas paginas e toastr

+++ app/Http/Controllers/AdminController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
+
 use App\Models\User;
 
 class AdminController extends Controller
@@ -12,15 +13,33 @@ class AdminController extends Controller
 
     public function userManagement()
     {
-        $users = User::select('id', 'name', 'email')->get();
-        
+        $users = User::select('id', 'name', 'email', 'usertype')->paginate(10);
+
         return view('admin.user-management', compact('users'));
     }
-    
+
+    public function user_search(Request $request)
+    {
+        $search = $request->input('search');
+        
+        $users = User::where('name', 'LIKE', '%' . $search . '%')->orWhere('usertype', 'LIKE', '%' . $search . '%')->paginate(3);
+
+        return view('admin.user-management', compact('users'));
+    }
+
     public function show($id)
     {
         $user = User::findOrFail($id);
         return view('users.show', compact('user'));
+    }
+    
+    public function support(){
+        return view('admin.support');
+    }
+
+    public function maintenance(){
+        
+        return view('admin.maintenance');
     }
 
     // Função para editar um utilizador
@@ -34,8 +53,14 @@ class AdminController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
+
+        if ($user->usertype === 'owner') {
+            toastr()->error('O Owner nao pode ser apagado');
+            return redirect()->back();
+        }
         $user->delete();
 
-        return redirect()->route('admin.user-management')->with('success', 'Utilizador apagado com sucesso.');
+        toastr()->success('Utilizador apagado com sucesso');
+        return redirect()->route('admin.user-management');
     }
 }
